# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Covid-19 REST API"
  version: "1.0.0"
  title: "Swagger Covid-19"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
tags:
  - name: "covid-19"
    description: "Everything about Covid-19"
    externalDocs:
      description: "Find out more"
      url: "http://swagger.io"
schemes:
  - "http"
paths:
  /confirmed:
    get:
      tags:
        - "confirmed"
      summary: "Get Confirmed Cases"
      description: ""
      operationId: "getConfirmed"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Covid"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "Covid"
      x-is-post-method: true
    post:
      tags:
        - "confirmed by country"
      summary: "Find covid by country"
      description: "Returns a single country confirmed cases"
      operationId: "getConfirmedByCountry"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Pet object that needs to be added to the store"
          required: true
          schema:
            type: object
            properties:
              country:
                type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Covid"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Covid not found"
      x-swagger-router-controller: "Covid"
      x-is-get-method: true
  /recovered:
    get:
      tags:
        - "recovered"
      summary: "Get Recovered Cases"
      description: ""
      operationId: "getRecovered"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Covid"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "Covid"
      x-is-post-method: true
    post:
      tags:
        - "recovered by country"
      summary: "Find covid by country"
      description: "Returns a single country recovered cases"
      operationId: "getRecoveredByCountry"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Pet object that needs to be added to the store"
          required: true
          schema:
            type: object
            properties:
              country:
                type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Covid"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Covid not found"
      x-swagger-router-controller: "Covid"
      x-is-get-method: true
  /deaths:
    get:
      tags:
        - "deaths"
      summary: "Get Death Cases"
      description: ""
      operationId: "getDeaths"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Covid"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "Covid"
      x-is-post-method: true
    post:
      tags:
        - "deaths by country"
      summary: "Find covid by country"
      description: "Returns a single country death cases"
      operationId: "getDeathsByCountry"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Pet object that needs to be added to the store"
          required: true
          schema:
            type: object
            properties:
              country:
                type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Covid"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Covid not found"
      x-swagger-router-controller: "Covid"
      x-is-get-method: true
  /latest:
    get:
      tags:
        - "latest"
      summary: "Get Lastest Cases"
      description: ""
      operationId: "getLatest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Covid"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "Covid"
      x-is-post-method: true
    post:
      tags:
        - "latest by country"
      summary: "Find covid by country"
      description: "Returns a single country latest cases"
      operationId: "getLatestByCountry"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Pet object that needs to be added to the store"
          required: true
          schema:
            type: object
            properties:
              country:
                type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Covid"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Covid not found"
      x-swagger-router-controller: "Covid"
      x-is-get-method: true
definitions:
  Covid:
    type: "object"
    properties:
      results:
        type: array
        description: "results for Covid-19 cases"
        items:
          type: object
          properties:
            "Province/State":
              type: string
            Lat:
              type: integer
            Long:
              type: integer
    title: "Covid-19 Cases"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"